refactor(pkg): Rename `kraft pkg show` to `kraft pkg info`
